// src/utils/enums/EnumProductSubGroup.ts
/* eslint-disable no-shadow */

// When updating this enum, you must also update the respective Enum in the backend project.
enum EnumProductSubGroup {
  // Employees = 'Employees',
  // Management = 'Management',
  // Corporate = 'Corporate',
  // Retail = 'Retail',
  // Wealthy = 'Wealthy',

  Employees = 'MitarbeiterInnen',
  Management = 'Management',
  Corporate = 'Firmenkunde / B2B',
  Retail = 'Privatkunde / B2C',
  Wealthy = 'vermögende Privatkunden',
}

export default EnumProductSubGroup;